+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Master
+
+- **BREAKING** The public interface of the `minim` module has changed significantly. List of changes:
+
+  - Removed `toCompactRefract` and `fromCompactRefract`
+
 # 0.13.0 - 2015-12-03
 
 - Added support for hyperlinks per [RFC 0008](https://github.com/refractproject/rfcs/blob/b6e390f7bbc960808ba053e172cccd9e4a81a04a/text/0008-add-hyperlinks.md)